backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.8.0
cryptography==42.0.5
Flask==3.0.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
google-auth==2.29.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
mysqlclient==2.2.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyMySQL==1.1.0
python-dotenv==1.0.1
python-tds==1.15.0
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.18.1
